refactor(embedded): change the return of `read()` in `ITransport` to `uint8_t`

+++ embedded/ble_transport.cpp
@@ -51,7 +51,7 @@ int BLETransport::available() {
     return rxBuffer.length();
 }
 
-int BLETransport::read() {
+uint8_t BLETransport::read() {
     if (rxBuffer.isEmpty()) {
         return -1;
     }
